bit fix on all operations

+++ src/Instruction.java
@@ -1,9 +1,9 @@
 public class Instruction {
     private final String opCode; //TODO: could be made enum
-    private final int addressA;
-    private final int addressB;
+    private final long addressA;
+    private final long addressB;
 
-    public Instruction(String opCode, int addressA, int addressB) {
+    public Instruction(String opCode, long addressA, long addressB) {
         this.opCode = opCode;
         this.addressA = addressA;
         this.addressB = addressB;
@@ -13,11 +13,11 @@ public class Instruction {
         return this.opCode;
     }
 
-    public int getAddressA() {
+    public long getAddressA() {
         return this.addressA;
     }
 
-    public int getAddressB() {
+    public long getAddressB() {
         return this.addressB;
     }
