+++ src/PCOperator.java
@@ -3,5 +3,5 @@ public interface PCOperator extends Operator {
     public int getNumB();
     public void setNumA(int numA);
     public void setNumB(int numB);
-    public int solve(boolean immediate, int pCounter); //for BZJ ad BZJi
+    public long solve(boolean immediate, long pCounter); //for BZJ ad BZJi
 }
